[PPS - 7] : Refactored color generator function for Board-Item and Task Item | Added Seed data | Refactored create task form.
