+++ src/features/boards/utils.ts
@@ -1,27 +1,11 @@
-const tailwindColors = [
-  "bg-red-500",
-  "bg-blue-500",
-  "bg-green-500",
-  "bg-yellow-500",
-  "bg-purple-500",
-  "bg-pink-500",
-  "bg-indigo-500",
-  "bg-teal-500",
-  "bg-orange-500",
-];
+export const getRandomColor = (boardId: string): string => {
+  // Generate a deterministic hash from boardId
+  const hash = Array.from(boardId).reduce((acc, char) => {
+    return char.charCodeAt(0) + ((acc << 5) - acc);
+  }, 0);
 
-const usedColors: Set<string> = new Set();
+  // Convert hash to hex color
+  const color = (hash & 0x00ffffff).toString(16).padStart(6, "0");
 
-export const getRandomColor = (): string => {
-  if (usedColors.size >= tailwindColors.length) {
-    usedColors.clear(); // Reset if all colors are used
-  }
-
-  let color: string;
-  do {
-    color = tailwindColors[Math.floor(Math.random() * tailwindColors.length)];
-  } while (usedColors.has(color));
-
-  usedColors.add(color);
-  return color;
+  return `#${color}`;
 };
