+++ src/features/tasks/api/use-copy-task.ts
@@ -1,6 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { client } from "@/lib/hono";
 import { InferRequestType, InferResponseType } from "hono";
+import { useGetBoardId } from "@/hooks/use-get-board-id";
 
 type RequestType = InferRequestType<
   (typeof client.api.tasks)["copy"][":columnId"][":id"]["$post"]
@@ -13,6 +14,7 @@ type ResponseType = InferResponseType<
 
 export const useCopyTask = () => {
   const queryClient = useQueryClient();
+  const boardId = useGetBoardId();
   const mutation = useMutation<ResponseType, Error, RequestType>({
     mutationFn: async ({ param, json }) => {
       const response = await client.api.tasks["copy"][":columnId"][":id"].$post(
@@ -31,6 +33,9 @@ export const useCopyTask = () => {
     onSuccess: ({ data }) => {
       queryClient.invalidateQueries({
         queryKey: ["tasks", data.columnId],
+      });
+      queryClient.invalidateQueries({
+        queryKey: ["columns-with-tasks", { boardId }],
       });
     },
   });
