+++ src/features/tasks/store/use-create-task-modal.ts
@@ -1,18 +1,15 @@
-import { useQueryState, parseAsBoolean } from "nuqs";
+import { useQueryState, parseAsString } from "nuqs";
 
 export const useCreateTaskModal = () => {
-  const [isOpen, setIsOpen] = useQueryState(
-    "create-task",
-    parseAsBoolean.withDefault(false).withOptions({ clearOnDefault: true })
-  );
+  const [id, setId] = useQueryState("create-task", parseAsString);
 
-  const open = () => setIsOpen(true);
-  const close = () => setIsOpen(false);
+  const open = (id: string) => setId(id);
+  const close = () => setId(null);
 
   return {
-    isOpen,
+    id,
     open,
     close,
-    setIsOpen,
+    setId,
   };
 };
