+++ src/features/tasks/utils.ts
@@ -1,99 +1,11 @@
-const tailwindBorderColors = [
-  "border-red-50",
-  "border-red-100",
-  "border-red-200",
-  "border-red-300",
-  "border-red-400",
-  "border-red-500",
-  "border-red-600",
-  "border-red-700",
-  "border-red-800",
-  "border-red-900",
-  "border-blue-50",
-  "border-blue-100",
-  "border-blue-200",
-  "border-blue-300",
-  "border-blue-400",
-  "border-blue-500",
-  "border-blue-600",
-  "border-blue-700",
-  "border-blue-800",
-  "border-blue-900",
-  "border-green-50",
-  "border-green-100",
-  "border-green-200",
-  "border-green-300",
-  "border-green-400",
-  "border-green-500",
-  "border-green-600",
-  "border-green-700",
-  "border-green-800",
-  "border-green-900",
-  "border-yellow-50",
-  "border-yellow-100",
-  "border-yellow-200",
-  "border-yellow-300",
-  "border-yellow-400",
-  "border-yellow-500",
-  "border-yellow-600",
-  "border-yellow-700",
-  "border-yellow-800",
-  "border-yellow-900",
-  "border-purple-50",
-  "border-purple-100",
-  "border-purple-200",
-  "border-purple-300",
-  "border-purple-400",
-  "border-purple-500",
-  "border-purple-600",
-  "border-purple-700",
-  "border-purple-800",
-  "border-purple-900",
-  "border-pink-50",
-  "border-pink-100",
-  "border-pink-200",
-  "border-pink-300",
-  "border-pink-400",
-  "border-pink-500",
-  "border-pink-600",
-  "border-pink-700",
-  "border-pink-800",
-  "border-pink-900",
-  "border-indigo-50",
-  "border-indigo-100",
-  "border-indigo-200",
-  "border-indigo-300",
-  "border-indigo-400",
-  "border-indigo-500",
-  "border-indigo-600",
-  "border-indigo-700",
-  "border-indigo-800",
-  "border-indigo-900",
-  "border-teal-50",
-  "border-teal-100",
-  "border-teal-200",
-  "border-teal-300",
-  "border-teal-400",
-  "border-teal-500",
-  "border-teal-600",
-  "border-teal-700",
-  "border-teal-800",
-  "border-teal-900",
-  "border-orange-50",
-  "border-orange-100",
-  "border-orange-200",
-  "border-orange-300",
-  "border-orange-400",
-  "border-orange-500",
-  "border-orange-600",
-  "border-orange-700",
-  "border-orange-800",
-  "border-orange-900",
-];
+export const getTaskColor = (position: number): string => {
+  // Generate a deterministic hash from position number
+  const hash = Math.abs(position * 1234567) % 0xffffff;
 
-export const GetRandomBorderColor = (): string => {
-  const randomIndex = Math.floor(Math.random() * tailwindBorderColors.length);
-  return tailwindBorderColors[randomIndex];
+  // Convert hash to hex color
+  const color = hash.toString(16).padStart(6, "0");
+
+  return `#${color}`;
 };
 
 export const CapitalizeFirstLetter = (word: string): string => {
